requirements.txt
streamlit==1.49.1
pypdf==6.8.0
python-docx==1.2.0
